Extract base class for user configuration handling
