+++ src/Administration/Resources/app/administration/src/module/sw-sales-channel/service/sales-channel-favorites.service.ts
@@ -1,70 +1,12 @@
 import Vue from 'vue';
+import UserConfigClass from '../../../core/service/support/user-config.class';
 
-const { Application, Service, Context, State } = Shopware;
-const { Criteria } = Shopware.Data;
+const { Application } = Shopware;
 
-enum USER_CONFIG_PERMISSIONS {
-    READ = 'user_config:read',
-    CREATE = 'user_config:create',
-    UPDATE = 'user_config:update'
-}
-
-interface CurrentUserObject {
-    id: string;
-    [index: string]: unknown;
-}
-
-interface UserConfigObject {
-    extensions: Record<string, unknown>;
-    id: string;
-    userId: string;
-    key: string;
-    value: string[];
-    isNew: () => boolean;
-}
-
-interface AclServiceInterface {
-    can(permission: string): boolean;
-}
-
-interface CriteriaInterface {
-    addFilter(filter: unknown): CriteriaInterface;
-    equals(field: string, value: string|number|boolean|null): unknown;
-}
-
-interface RepositorySearchResultInterface {
-    total: number;
-    first(): unknown;
-}
-
-interface RepositoryInterface {
-    create(context: unknown): unknown;
-    search(criteria: CriteriaInterface, context: unknown): Promise<RepositorySearchResultInterface>;
-    save(entity: unknown, context: unknown): Promise<void>;
-}
-
-class SalesChannelFavoritesService {
+class SalesChannelFavoritesService extends UserConfigClass {
     static USER_CONFIG_KEY = 'sales-channel-favorites';
 
-    private userConfigRepository = <RepositoryInterface><unknown> Service('repositoryFactory').create('user_config');
-
-    private currentUserId = this.getCurrentUserId();
-
-    private userConfig = this.createUserConfigEntity();
-
     private state: { favorites: string[] } = Vue.observable({ favorites: [] });
-
-    private aclService = <AclServiceInterface> Service('acl');
-
-    constructor() {
-        void this.initService();
-    }
-
-    private async initService(): Promise<void> {
-        this.userConfig = await this.getUserConfig();
-
-        this.state.favorites = this.userConfig.value;
-    }
 
     public getFavoriteIds(): string[] {
         return this.state.favorites;
@@ -86,64 +28,17 @@ class SalesChannelFavoritesService {
         void this.saveUserConfig();
     }
 
-    public refresh(): void {
-        this.userConfig = this.createUserConfigEntity();
-        void this.initService();
+    protected getConfigurationKey(): string {
+        return SalesChannelFavoritesService.USER_CONFIG_KEY;
     }
 
-    private async getUserConfig(): Promise<UserConfigObject> {
-        if (!this.aclService.can(USER_CONFIG_PERMISSIONS.READ)) {
-            void Promise.resolve(this.userConfig);
-        }
-
-        const response = await this.userConfigRepository.search(this.getCriteria(), Context.api);
-        const userConfig = <UserConfigObject> (response.total ? response.first() : this.userConfig);
-
-        return this.handleEmptyUserConfig(userConfig);
+    protected async readUserConfig(): Promise<void> {
+        this.userConfig = await this.getUserConfig();
+        this.state.favorites = this.userConfig.value;
     }
 
-    private async saveUserConfig(): Promise<void> {
-        if (!this.aclService.can(USER_CONFIG_PERMISSIONS.CREATE) || !this.aclService.can(USER_CONFIG_PERMISSIONS.UPDATE)) {
-            void Promise.resolve();
-        }
-
+    protected setUserConfig(): void {
         this.userConfig.value = this.state.favorites;
-
-        await this.userConfigRepository.save(this.userConfig, Context.api);
-        await this.initService();
-    }
-
-    private createUserConfigEntity(): UserConfigObject {
-        const entity = <UserConfigObject> this.userConfigRepository.create(Context.api);
-
-        Object.assign(entity, {
-            userId: this.currentUserId,
-            key: SalesChannelFavoritesService.USER_CONFIG_KEY,
-            value: [],
-        });
-
-        return entity;
-    }
-
-    private handleEmptyUserConfig(userConfig: UserConfigObject): UserConfigObject {
-        if (!Array.isArray(userConfig.value)) {
-            userConfig.value = [];
-        }
-
-        return userConfig;
-    }
-
-    private getCriteria(): CriteriaInterface {
-        const criteria = <CriteriaInterface><unknown> new Criteria(1, 25);
-
-        criteria.addFilter(Criteria.equals('key', SalesChannelFavoritesService.USER_CONFIG_KEY));
-        criteria.addFilter(Criteria.equals('userId', this.currentUserId));
-
-        return criteria;
-    }
-
-    private getCurrentUserId(): string {
-        return (State.get('session').currentUser as CurrentUserObject).id;
     }
 }
